backend/requirements.txt
fastapi==0.105.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
ultraimport==0.0.7
python-jose[cryptography]==3.3.0